open64 merge r3610:3640.
See the open64 commit log for more details.


git-svn-id: https://svn.open64.net/svnroot/open64/branches/OpenUH@3641 ead328b1-e903-0410-a19d-eacee4b031f6

+++ osprey/be/opt/opt_proactive.h
@@ -437,7 +437,7 @@ protected:
   void Delete();
   void Set_cu(COMP_UNIT * i) { _cu = i; }
   COMP_UNIT * Get_cu();
-  WN * Get_const_wn(long long);
+  WN * Get_const_wn(INT64);
   WN * Get_cond(SC_NODE *, BOOL);
   WN * Merge_cond(WN *, WN *, OPERATOR);
   BOOL Can_reorder_cond(WN *, WN *);
